+++ src/main/java/org/plumelib/util/CollectionsPlume.java
@@ -55,22 +55,7 @@ public final class CollectionsPlume {
   ///
 
   /**
-   * Returns the sorted version of the list. Does not alter the list. Simply calls {@code
-   * Collections.sort(List<T>, Comparator<? super T>)}.
-   *
-   * @return a sorted version of the list
-   * @param <T> type of elements of the list
-   * @param l a list to sort
-   * @param c a sorted version of the list
-   */
-  public static <T> List<T> sortList(List<T> l, Comparator<@MustCallUnknown ? super T> c) {
-    List<T> result = new ArrayList<>(l);
-    Collections.sort(result, c);
-    return result;
-  }
-
-  /**
-   * Returns true iff the list does not contain duplicate elements.
+   * Returns true iff the list does not contain duplicate elements, according to {@code equals()}.
    *
    * <p>The implementation uses O(n) time and O(n) space.
    *
@@ -100,7 +85,7 @@ public final class CollectionsPlume {
   }
 
   /**
-   * Returns true iff the list does not contain duplicate elements.
+   * Returns true iff the list does not contain duplicate elements, according to {@code equals()}.
    *
    * <p>The implementation uses O(n) time and O(n) space.
    *
@@ -109,13 +94,32 @@ public final class CollectionsPlume {
    * @return true iff a does not contain duplicate elements
    */
   @Pure
-  public static <T> boolean noDuplicates(List<T> a) {
+  public static <T> boolean hasNoDuplicates(List<T> a) {
     return !hasDuplicates(a);
   }
 
   /**
-   * Returns a copy of the list (never the original list) with duplicates removed, but retaining the
-   * original order.
+   * Returns true iff the list does not contain duplicate elements, according to {@code equals()}.
+   *
+   * <p>The implementation uses O(n) time and O(n) space.
+   *
+   * @param <T> the type of the elements
+   * @param a a list
+   * @return true iff a does not contain duplicate elements
+   * @deprecated use {@link #hasNoDuplicates(List)}
+   */
+  @Deprecated // 2023-11-30
+  // @InlineMe(
+  //     replacement = "CollectionsPlume.hasNoDuplicates(a)",
+  //     imports = "org.plumelib.util.CollectionsPlume")
+  @Pure
+  public static <T> boolean noDuplicates(List<T> a) {
+    return hasNoDuplicates(a);
+  }
+
+  /**
+   * Returns a copy of the list (never the original list) with duplicates (according to {@code
+   * equals()}) removed, but retaining the original order. The argument is not modified.
    *
    * @param <T> type of elements of the list
    * @param l a list to remove duplicates from
@@ -130,9 +134,9 @@ public final class CollectionsPlume {
   }
 
   /**
-   * Returns a copy of the list with duplicates removed, but retaining the original order. May
-   * return its argument if its argument has no duplicates, but is not guaranteed to do so. The
-   * argument is not modified.
+   * Returns a copy of the list with duplicates (according to {@code equals()}) removed, but
+   * retaining the original order. May return its argument if its argument has no duplicates, but is
+   * not guaranteed to do so. The argument is not modified.
    *
    * <p>If the element type implements {@link Comparable}, use {@link #withoutDuplicatesSorted} or
    * {@link #withoutDuplicatesComparable}.
@@ -151,9 +155,9 @@ public final class CollectionsPlume {
   }
 
   /**
-   * Returns a list with the same contents as its argument, but sorted and without duplicates. May
-   * return its argument if its argument is sorted and has no duplicates, but is not guaranteed to
-   * do so. The argument is not modified.
+   * Returns a list with the same contents as its argument, but sorted and without duplicates
+   * (according to {@code equals()}). May return its argument if its argument is sorted and has no
+   * duplicates, but is not guaranteed to do so. The argument is not modified.
    *
    * <p>This is like {@link #withoutDuplicates}, but requires the list elements to implement {@link
    * Comparable}, and thus can be more efficient.
@@ -202,6 +206,22 @@ public final class CollectionsPlume {
     } else {
       return new ArrayList<>(set);
     }
+  }
+
+  /**
+   * Returns the sorted version of the list. Does not alter the list. Simply calls {@code
+   * Collections.sort(List<T>, Comparator<? super T>)} on a copy.
+   *
+   * @return a sorted version of the list
+   * @param <T> type of elements of the list
+   * @param l a list to sort; is not side-effected
+   * @param c a sorted version of the list
+   */
+  // TODO: rename to "sorted()".
+  public static <T> List<T> sortList(List<T> l, Comparator<@MustCallUnknown ? super T> c) {
+    List<T> result = new ArrayList<>(l);
+    Collections.sort(result, c);
+    return result;
   }
 
   /**
@@ -282,7 +302,6 @@ public final class CollectionsPlume {
    * @return the elements (once each) that appear more than once in the given collection
    */
   public static <T> Collection<T> duplicates(Collection<T> c) {
-    // Inefficient (because of streams) but simple implementation.
     Set<T> withoutDuplicates = new HashSet<>();
     Set<T> duplicates = new LinkedHashSet<>();
     for (T elt : c) {
@@ -451,6 +470,8 @@ public final class CollectionsPlume {
    *
    * <pre>import static org.plumelib.util.CollectionsPlume.mapList;</pre>
    *
+   * This method is just like {@link #transform}, but with the arguments in the other order.
+   *
    * @param <FROM> the type of elements of the given array
    * @param <TO> the type of elements of the result list
    * @param f a function
@@ -555,7 +576,33 @@ public final class CollectionsPlume {
 
   /**
    * Returns a new list containing only the elements for which the filter returns true. To modify
-   * the collection in place, use {@code Collection#removeIf}.
+   * the collection in place, use {@code Collection#removeIf} instead of this method.
+   *
+   * <p>Using streams gives an equivalent list but is less efficient and more verbose:
+   *
+   * <pre>{@code
+   * coll.stream().filter(filter).collect(Collectors.toList());
+   * }</pre>
+   *
+   * @param <T> the type of elements
+   * @param coll a collection
+   * @param filter a predicate
+   * @return a new list with the elements for which the filter returns true
+   * @deprecated use {@link #filter} instead
+   */
+  @Deprecated // 2023-11-30
+  // @InlineMe(
+  //     replacement = "CollectionsPlume.filter(coll, filter)",
+  //     imports = "org.plumelib.util.CollectionsPlume")
+  public static <T> List<T> listFilter(Iterable<T> coll, Predicate<? super T> filter) {
+    return filter(coll, filter);
+  }
+
+  // TODO: This should return a collection of the same type as the input.  Currently it always
+  // returns a list.
+  /**
+   * Returns a new list containing only the elements for which the filter returns true. To modify
+   * the collection in place, use {@code Collection#removeIf} instead of this method.
    *
    * <p>Using streams gives an equivalent list but is less efficient and more verbose:
    *
@@ -568,7 +615,7 @@ public final class CollectionsPlume {
    * @param filter a predicate
    * @return a new list with the elements for which the filter returns true
    */
-  public static <T> List<T> listFilter(Collection<T> coll, Predicate<? super T> filter) {
+  public static <T> List<T> filter(Iterable<T> coll, Predicate<? super T> filter) {
     List<T> result = new ArrayList<>();
     for (T elt : coll) {
       if (filter.test(elt)) {
@@ -576,6 +623,214 @@ public final class CollectionsPlume {
       }
     }
     return result;
+  }
+
+  /**
+   * Returns true if any element of the collection matches the predicate.
+   *
+   * <p>Using streams gives an equivalent result but is less efficient:
+   *
+   * <pre>{@code
+   * coll.stream().anyMatch(predicate);
+   * }</pre>
+   *
+   * @param <T> the type of elements
+   * @param coll a collection
+   * @param predicate a non-interfering, stateless predicate
+   * @return true if any element of the collection matches the predicate
+   */
+  public static <T> boolean anyMatch(Iterable<T> coll, Predicate<? super T> predicate) {
+    for (T elt : coll) {
+      if (predicate.test(elt)) {
+        return true;
+      }
+    }
+    return false;
+  }
+
+  /**
+   * Returns true if all elements of the collection match the predicate.
+   *
+   * <p>Using streams gives an equivalent result but is less efficient:
+   *
+   * <pre>{@code
+   * coll.stream().allMatch(predicate);
+   * }</pre>
+   *
+   * @param <T> the type of elements
+   * @param coll a collection
+   * @param predicate a non-interfering, stateless predicate
+   * @return true if all elements of the collection match the predicate
+   */
+  public static <T> boolean allMatch(Iterable<T> coll, Predicate<? super T> predicate) {
+    for (T elt : coll) {
+      if (!predicate.test(elt)) {
+        return false;
+      }
+    }
+    return true;
+  }
+
+  /**
+   * Returns true if no element of the collection matches the predicate.
+   *
+   * <p>Using streams gives an equivalent result but is less efficient:
+   *
+   * <pre>{@code
+   * coll.stream().noneMatch(predicate);
+   * }</pre>
+   *
+   * @param <T> the type of elements
+   * @param coll a collection
+   * @param predicate a non-interfering, stateless predicate
+   * @return true if no element of the collection matches the predicate
+   */
+  public static <T> boolean noneMatch(Iterable<T> coll, Predicate<? super T> predicate) {
+    for (T elt : coll) {
+      if (predicate.test(elt)) {
+        return false;
+      }
+    }
+    return true;
+  }
+
+  /**
+   * Returns the first index of the given value in the list, starting at the given index. Uses
+   * {@code Object.equals()} for comparison.
+   *
+   * @param list a list
+   * @param value the value to search for
+   * @param start the starting index
+   * @return the index of the value in the list, at or after the given index
+   */
+  public static int indexOf(List<?> list, Object value, int start) {
+    int idx = list.subList(start, list.size()).indexOf(value);
+    return idx == -1 ? -1 : idx + start;
+  }
+
+  /**
+   * Represents a replacement of one range of a collection by another collection.
+   *
+   * @param <T> the type of collection elements
+   */
+  public static class Replacement<T> {
+    /** The first line to replace, inclusive. */
+    public final int start;
+
+    /** The last line to replace, <em>inclusive</em>. May be equal to {@code start}-1. */
+    public final int end;
+
+    /** The new (replacement) elements. */
+    final Collection<T> elements;
+
+    /**
+     * Creates a new Replacement.
+     *
+     * @param start the first line to replace, inclusive
+     * @param end the last line to replace, exclusive
+     * @param elements the new (replacement) elements
+     */
+    private Replacement(int start, int end, Collection<T> elements) {
+      this.start = start;
+      this.end = end;
+      this.elements = elements;
+      if (end < start - 1) {
+        throw new Error("Invalid <start,end> pair: " + this);
+      }
+    }
+
+    /**
+     * Creates a new Replacement.
+     *
+     * @param <T> the type of elements of the list
+     * @param start the first line to replace, inclusive
+     * @param end the last line to replace, exclusive
+     * @param elements the new (replacement) elements
+     * @return a new Replacement
+     */
+    public static <T> Replacement<T> of(int start, int end, Collection<T> elements) {
+      return new Replacement<T>(start, end, elements);
+    }
+
+    @Override
+    public String toString(@GuardSatisfied Replacement<T> this) {
+      return "Replacement{" + start + ", " + end + ", " + elements + "}";
+    }
+  }
+
+  /**
+   * Performs a set of replacements on the given collection, returning the transformed result (as a
+   * list).
+   *
+   * @param <T> the type of collection elements
+   * @param c a collection
+   * @param replacements the replacements to perform on the collection, in order from the beginning
+   *     of the collection to the end
+   * @return the transformed collection, as a new list (even if no changes were made)
+   */
+  public static <T> List<T> replace(Iterable<T> c, Iterable<Replacement<T>> replacements) {
+    List<T> result = new ArrayList<>();
+    Iterator<T> cItor = c.iterator();
+    int cIndex = -1; // the index into c
+    Iterator<Replacement<T>> replacementItor = replacements.iterator();
+    while (replacementItor.hasNext()) {
+      Replacement<T> replacement = replacementItor.next();
+      while (cIndex < replacement.start - 1) {
+        result.add(cItor.next());
+        cIndex++;
+      }
+      result.addAll(replacement.elements);
+      while (cIndex < replacement.end) {
+        cItor.next();
+        cIndex++;
+      }
+    }
+    while (cItor.hasNext()) {
+      result.add(cItor.next());
+    }
+    return result;
+  }
+
+  /**
+   * Performs a set of replacements on the given array, returning the transformed result (as a
+   * list).
+   *
+   * @param <T> the type of collection elements
+   * @param c an array
+   * @param replacements the replacements to perform on the arary, in order from the beginning of
+   *     the list to the end
+   * @return the transformed collection, as a list
+   */
+  public static <T> List<T> replace(T[] c, Collection<Replacement<T>> replacements) {
+    return replace(Arrays.asList(c), replacements);
+  }
+
+  /**
+   * Returns true if the second list is a subsequence (not necessarily contiguous) of the first.
+   *
+   * @param <T> the type of elements of the list
+   * @param longer a list
+   * @param shorter a list
+   * @return true if the second list is a subsequence (not necessarily contiguous) of the first
+   */
+  // TODO: This could take as input a RandomAccess.
+  @SuppressWarnings("signedness")
+  public static <T> boolean isSubsequenceMaybeNonContiguous(
+      Iterable<T> longer, Iterable<T> shorter) {
+    Iterator<T> itorLonger = longer.iterator();
+    Iterator<T> itorShorter = shorter.iterator();
+    outerLoop:
+    while (itorShorter.hasNext()) {
+      T eltShorter = itorShorter.next();
+      while (itorLonger.hasNext()) {
+        T eltLonger = itorLonger.next();
+        if (Objects.equals(eltShorter, eltLonger)) {
+          continue outerLoop;
+        }
+      }
+      return false;
+    }
+    return true;
   }
 
   ///////////////////////////////////////////////////////////////////////////
@@ -1092,7 +1347,7 @@ public final class CollectionsPlume {
   }
 
   /**
-   * An iterator that only returns elements that match the given Filter.
+   * An iterator that only returns elements that match the given predicate.
    *
    * @param <T> the type of elements of the iterator
    */
@@ -1101,17 +1356,17 @@ public final class CollectionsPlume {
     Iterator<T> itor;
 
     /** The predicate that determines which elements to retain. */
-    Filter<T> filter;
+    Predicate<T> predicate;
 
     /**
-     * Create an iterator that only returns elements of {@code itor} that match the given Filter.
+     * Create an iterator that only returns elements of {@code itor} that match the given predicate.
      *
      * @param itor the Iterator to filter
-     * @param filter the predicate that determines which elements to retain
+     * @param predicate the predicate that determines which elements to retain
      */
-    public FilteredIterator(Iterator<T> itor, Filter<T> filter) {
+    public FilteredIterator(Iterator<T> itor, Predicate<T> predicate) {
       this.itor = itor;
-      this.filter = filter;
+      this.predicate = predicate;
     }
 
     /** A marker object, distinct from any object that the iterator can return. */
@@ -1135,7 +1390,7 @@ public final class CollectionsPlume {
     public boolean hasNext(@GuardSatisfied FilteredIterator<T> this) {
       while (!currentValid && itor.hasNext()) {
         current = itor.next();
-        currentValid = filter.accept(current);
+        currentValid = predicate.test(current);
       }
       return currentValid;
     }