+++ src/main/java/org/plumelib/util/package-info.java
@@ -65,8 +65,8 @@
  *       regular expression.
  *   <dt>{@link org.plumelib.util.StringsPlume}
  *   <dd>Utility methods that manipulate Strings: replacement; prefixing and indentation; splitting
- *       and joining; quoting and escaping; whitespace; comparisons; StringTokenizer; debugging
- *       variants of toString; diagnostic output; miscellaneous.
+ *       and joining; quoting and escaping; whitespace; comparisons; StringTokenizer; version
+ *       numbers; debugging variants of toString; diagnostic output; miscellaneous.
  *       <!--
  *   <dt>{link org.plumelib.util.CountingPrintWriter CountingPrintWriter}
  *   <dd>Prints formatted representations of objects to a text-output stream counting the number of
@@ -123,9 +123,6 @@
  * <h3 id="interfaces">Utility interfaces</h3>
  *
  * <dl>
- *   <dt>{@link org.plumelib.util.Filter Filter}
- *   <dd>Interface for things that make boolean decisions. This is inspired by {@code
- *       java.io.FilenameFilter}.
  *   <dt>{@link org.plumelib.util.Partitioner Partitioner}
  *   <dd>A Partitioner accepts Objects and assigns them to an equivalence class.
  * </dl>